+++ Source/WebKit/SourcesGTK.txt
@@ -249,6 +249,7 @@ UIProcess/Notifications/glib/NotificationService.cpp
 
 UIProcess/linux/MemoryPressureMonitor.cpp
 
+UIProcess/WebsiteData/glib/WebsiteDataStoreGLib.cpp
 UIProcess/WebsiteData/soup/WebsiteDataStoreSoup.cpp
 UIProcess/WebsiteData/unix/WebsiteDataStoreUnix.cpp
 
@@ -257,7 +258,6 @@ UIProcess/cairo/BackingStoreCairo.cpp @no-unify
 UIProcess/glib/WebPageProxyGLib.cpp
 UIProcess/glib/WebProcessPoolGLib.cpp
 UIProcess/glib/WebProcessProxyGLib.cpp
-UIProcess/glib/WebsiteDataStoreGLib.cpp @no-unify
 
 UIProcess/gstreamer/WebPageProxyGStreamer.cpp
