+++ Source/WebKit/UIProcess/WebsiteData/WebsiteDataStoreConfiguration.cpp
@@ -91,9 +91,7 @@ void WebsiteDataStoreConfiguration::initializePaths()
     setCacheStorageDirectory(WebsiteDataStore::defaultCacheStorageDirectory(m_baseCacheDirectory));
     setNetworkCacheDirectory(WebsiteDataStore::defaultNetworkCacheDirectory(m_baseCacheDirectory));
     setMediaCacheDirectory(WebsiteDataStore::defaultMediaCacheDirectory(m_baseCacheDirectory));
-#if USE(GLIB) || PLATFORM(COCOA)
     setHSTSStorageDirectory(WebsiteDataStore::defaultHSTSStorageDirectory(m_baseCacheDirectory));
-#endif
 #if ENABLE(ARKIT_INLINE_PREVIEW)
     setModelElementCacheDirectory(WebsiteDataStore::defaultModelElementCacheDirectory());
 #endif
